cast int to unsigned

+++ Premiere/EncodingData.h
@@ -9,7 +9,17 @@ typedef struct EncodingData
 	int planes;
 	const char *pix_fmt;
 	char *plane[8] = {};
-	unsigned int stride[8] = { -1, -1, -1, -1, -1, -1, -1, -1 };
+	unsigned int stride[8] = 
+	{ 
+		static_cast<unsigned>(-1),
+		static_cast<unsigned>(-1),
+		static_cast<unsigned>(-1),
+		static_cast<unsigned>(-1),
+		static_cast<unsigned>(-1),
+		static_cast<unsigned>(-1),
+		static_cast<unsigned>(-1),
+		static_cast<unsigned>(-1)
+	};
 	int pass;
 	struct filters
 	{
